Setup a scheduled runnable that pings each memcached node configured to see if the connection is still alive.
The PingServer class finds keys for each server,then stores them upon construction.

When the runnable executes, it iterates through the 'available' connections, and checks to see if they are alive.  If they aren't alive, it attempts to connect until either its alive, or it gets evicted from the 'available' connections